backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""

authors = [
    "Innovatica LTD",
]

repository = ""
homepage = ""

# README file(s) are used as the package description
readme = ["README.md", ]

# Keywords (translated to tags on the package index)
keywords = [ "backend"]

[tool.poetry.dependencies]
# Compatible Python versions
python = ">=3.12,<4"
django = "~=5.1.3"
django-environ = "~=0.10.0"
django-storages = "~=1.13.2"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
waitress = "~=2.1.2"
whitenoise = "~=6.5.0"
djangorestframework = "~=3.14.0"
dj-rest-auth= "~=4.0.1"
djangorestframework-simplejwt= "~=5.2.2"
django-extensions = "~=3.2.3"
django-cors-headers = "~=4.1.0"
drf-extra-fields = "~=3.5.0"
django-nested-admin = "~=4.0.2"
drf-writable-nested = "~=0.7.0"
drf-nested-routers = "~=0.93.4"
django-model-utils = "~=4.3.1"
munch = "~=3.0.0"
sentry-sdk = "~=1.33.1"
drf-spectacular = "^0.26.3"
django-import-export = "~=3.2.0"
pillow = "~=9.5.0"
django-json-widget = "^1.1.1"
django-jazzmin = "^2.6.0"
celery = "^5.3.1"
websockets = "^11.0.3"
django-timezone-field = "^6.0.1"
boto3 = "^1.28.59"
django-colorfield = "^0.10.1"
channels = {extras = ["daphne"], version = "^4.0.0"}
channels-redis = "^4.2.0"
django-redis = "^5.4.0"
itsdangerous = "^2.1.2"
django-webshell = "^0.4.0"
django-user-agents = "^0.4.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.1.0"}


[tool.poetry.group.dev.dependencies]
poetrip = "^1.0.0"

[tool.chuy]
migrate = "python manage.py migrate"
mig = "python manage.py migrate"
mkmig = "python manage.py makemigrations"
mkmigmerge = "python manage.py makemigrations --merge"
sp = "python manage.py shell_plus"
